Added the calendar script. Cleanup.
